fix: handle no transactions, negative taxes owed as possible states on tax summary
