+++ src/components/TaxEstimatesSummaryCard/states/TaxEstimatesSummaryCardError.tsx
@@ -1,0 +1,15 @@
+import { useTranslation } from 'react-i18next'
+
+import { DataState, DataStateStatus } from '@components/DataState/DataState'
+
+export const TaxEstimatesSummaryCardError = () => {
+  const { t } = useTranslation()
+  return (
+    <DataState
+      className='Layer__TaxEstimatesSummaryCard__DataState Layer__TaxEstimatesSummaryCard__DataState--error Layer__data-state--reset'
+      status={DataStateStatus.info}
+      title={t('taxEstimates:error.load_tax_estimates_summary', 'We couldn\'t load your tax summary')}
+      description={t('taxEstimates:error.load_tax_estimates_summary_description', 'An error occurred while loading your tax summary. Please check your connection and try again.')}
+    />
+  )
+}
